Refactor: 역할에 따른 패키지 위치 수정
