+++ src/main/java/com/prgms/allen/dining/api/customer/restaurant/CustomerRestaurantApi.java
@@ -1,4 +1,4 @@
-package com.prgms.allen.dining.web.domain.customer.restaurant;
+package com.prgms.allen.dining.api.customer.restaurant;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.prgms.allen.dining.domain.restaurant.RestaurantService;
 import com.prgms.allen.dining.domain.restaurant.dto.MenuDetailRes;
-import com.prgms.allen.dining.domain.restaurant.dto.RestaurantSimpleRes;
 import com.prgms.allen.dining.domain.restaurant.dto.RestaurantDetailResForCustomer;
+import com.prgms.allen.dining.domain.restaurant.dto.RestaurantSimpleRes;
 
 @RestController
 @RequestMapping("/customer/api/restaurants")
