+++ src/main/java/com/prgms/allen/dining/domain/restaurant/RestaurantDuplicateCreationException.java
@@ -1,4 +1,4 @@
-package com.prgms.allen.dining.global.error.exception;
+package com.prgms.allen.dining.domain.restaurant;
 
 import com.prgms.allen.dining.global.error.ErrorCode;
