+++ src/main/java/com/prgms/allen/dining/web/interceptor/InterceptorConfig.java
@@ -1,13 +1,11 @@
-package com.prgms.allen.dining.web.config;
+package com.prgms.allen.dining.web.interceptor;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.prgms.allen.dining.web.interceptor.ApiLogInterceptor;
-
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
+public class InterceptorConfig implements WebMvcConfigurer {
 
 	@Override
 	public void addInterceptors(InterceptorRegistry registry) {
